Reorganise: move sources from /extension to /src.
/extension is to be a build output folder. Folder structure in /src
should match that of the output, so that e.g. a <script> tag's src
reference can be followed (except the jsx is renamed to js).

# Conflicts:
#	.gitignore
#	extension/page-viewer/localpage.html
#	gulpfile.babel.js
#	src/background.js
#	src/local-page/index.js
#	src/manifest.json
#	src/omnibar.js
#	src/overview/components/VisitAsListItem.jsx
#	src/popup.html
#	src/popup/popup.js
