+++ .gitignore
@@ -1,15 +1,5 @@
 node_modules
+extension
 dist
 .chrome-extension-key.pem
-extension/lib
-extension/browser-polyfill.js
-extension/background.js
-extension/content_script.js
-extension/overview/overview.js
-extension/overview/style.css
-extension/options/options.js
-extension/options/style.css
-extension/page-viewer/localpage.js
-extension/popup/popup.js
-extension/popup/style.css
 jest_0/
