--- extension/page-viewer/localpage.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8" />
-        <link rel="stylesheet" href="localpage.css" />
-    </head>
-    <body>
-        <div id="bar"></div>
-        <script src="/lib/browser-polyfill.js"></script>
-        <script src="localpage.js"></script>
-    </body>
-</html>
